rename the base rotation (deterministic/random hadamard) directory name.
