+++ auto_round/algorithms/transforms/hadamard/utils/math.py
