+++ auto_round/experimental/apply_rotation_transform.py
@@ -3,10 +3,10 @@
 """Backward-compat re-export shim.
 
 The canonical implementation now lives in
-:mod:`auto_round.algorithms.transforms.quarot.dispatcher`.
+:mod:`auto_round.algorithms.transforms.hadamard.dispatcher`.
 """
 
-from auto_round.algorithms.transforms.quarot.dispatcher import (  # noqa: F401
+from auto_round.algorithms.transforms.hadamard.dispatcher import (  # noqa: F401
     apply_hadamard_rotation,
     resolve_hadamard_backend,
 )
