+++ auto_round/experimental/utils.py
@@ -16,7 +16,7 @@ from typing import Any
 
 import torch
 
-from auto_round.algorithms.transforms.quarot import HADAMARDS, RotationConfig
+from auto_round.algorithms.transforms.hadamard import HADAMARDS, RotationConfig
 from auto_round.compressors.utils import is_mx_fp, is_nv_fp
 from auto_round.utils import logger
 
@@ -129,7 +129,7 @@ def is_triton_kernel_available(data_type: str) -> bool:
         return False
 
     try:
-        from auto_round.algorithms.transforms.quarot.utils.triton.mxfp4 import (  # pylint: disable=E0401
+        from auto_round.algorithms.transforms.hadamard.utils.triton.mxfp4 import (  # pylint: disable=E0401
             mxfp4_forward_kernel_wrapper,
         )
     except Exception:
@@ -139,19 +139,19 @@ def is_triton_kernel_available(data_type: str) -> bool:
 
 
 def dump_group_size_to_rotation_config(rotation_config: str | dict | RotationConfig, group_size: int):
-    from auto_round.algorithms.transforms.quarot.config import dump_group_size_to_rotation_config as _impl
+    from auto_round.algorithms.transforms.hadamard.config import dump_group_size_to_rotation_config as _impl
 
     return _impl(rotation_config, group_size)
 
 
 def to_dict_rotation_config(rotation_config: str | dict | RotationConfig):
-    from auto_round.algorithms.transforms.quarot.config import to_dict_rotation_config as _impl
+    from auto_round.algorithms.transforms.hadamard.config import to_dict_rotation_config as _impl
 
     return _impl(rotation_config)
 
 
 def normalize_rotation_config(rotation_config: str | dict | RotationConfig | None, data_type: str) -> dict[str, Any]:
-    from auto_round.algorithms.transforms.quarot.config import normalize_rotation_config as _impl
+    from auto_round.algorithms.transforms.hadamard.config import normalize_rotation_config as _impl
 
     return _impl(rotation_config, data_type)
